put PrimordialAccountDetails in lib.rs
